Refactor: TabBarIndicatorSizePopupMenu to use nullable enum TabBarIndicatorSize
The enum in the popup is as value and callback, default null value is first.

Has built in default label, but can be overridden for both default value label and disabled label.

This PopupMenu is an example of how the rest of them should be made. The usage is much simpler now.

Supports ListTileReveal as well.