Playground: Try SingleChildScrollView instead of ListView
Evaluating SingleChildScrollView vs ListView
The content is always fixed and known amount of widgets, but
always a single child, so a ListView is not needed, trying the
the SingleChildScrollView instead. Which for uses cases
with only a few items, or just one, like this case might be
faster. This test trail started Oct 12, 2024.